+++ src/BloomBrowserUI/bookEdit/readerSetup/readerSetup.ui.ts
@@ -213,9 +213,9 @@ function requestWordsForSelectedStage():void {
 
   var useSampleWords = $('input[name="words-or-letters"]:checked').val() === '1';
   if (useSampleWords)
-    accordionWindow().postMessage('Words\n' + (<HTMLTableCellElement>tr.cells[0]).innerHTML, '*');
+    toolboxWindow().postMessage('Words\n' + (<HTMLTableCellElement>tr.cells[0]).innerHTML, '*');
   else
-    accordionWindow().postMessage('Words\n' + knownGPCS, '*');
+    toolboxWindow().postMessage('Words\n' + knownGPCS, '*');
 }
 
 /**
@@ -471,10 +471,10 @@ function tabBeforeActivate(ui): void {
       previousMoreWords = (<HTMLInputElement>document.getElementById('dls_more_words')).value;
 
       // save the changes and update lists
-      var accordion = accordionWindow();
+      var toolbox = toolboxWindow();
       saveChangedSettings(function() {
-        if (typeof accordion['readerSampleFilesChanged'] === 'function')
-          accordion['readerSampleFilesChanged']();
+        if (typeof toolbox['readerSampleFilesChanged'] === 'function')
+          toolbox['readerSampleFilesChanged']();
       });
     }
   }
@@ -872,7 +872,7 @@ function setWordContainerHeight() {
 function finishInitializing() {
   $('#stages-table').find('tbody').sortable({ stop: updateStageNumbers });
   $('#levels-table').find('tbody').sortable({ stop: updateLevelNumbers });
-  accordionWindow().postMessage('Texts', '*');
+  toolboxWindow().postMessage('Texts', '*');
   setWordContainerHeight();
 }
 
@@ -881,17 +881,17 @@ function finishInitializing() {
  * has changed.
  */
 function wordListChangedCallback() {
-  var accordion = accordionWindow();
-  if (!accordion) return;
-  accordion.postMessage('Texts', '*');
+  var toolbox = toolboxWindow();
+  if (!toolbox) return;
+  toolbox.postMessage('Texts', '*');
   requestWordsForSelectedStage();
 }
 
 $(document).ready(function () {
   attachEventHandlers();
   $('body').find('*[data-i18n]').localize(finishInitializing);
-  var accordion = accordionWindow();
-  accordion['addWordListChangedListener']('wordListChanged.ReaderSetup', wordListChangedCallback);
+  var toolbox = toolboxWindow();
+  toolbox['addWordListChangedListener']('wordListChanged.ReaderSetup', wordListChangedCallback);
   // found solution to longpress access here:
   // http://stackoverflow.com/questions/3032770/execute-javascript-function-in-a-another-iframe-when-parent-is-from-different-do
   var pageIframe = parent.frames['page'];