+++ src/BloomBrowserUI/bookEdit/toolbox/talkingBook/talkingBook.ts
@@ -2,11 +2,11 @@
 
     // Running for real, and jquery properly loaded first
     $(document).ready(function () {
-        // We attach to the whole accordion, not to the #showRecordingTools directly,
+        // We attach to the whole toolbox, not to the #showRecordingTools directly,
         // because if the talking books tab is initially disabled, even after the page is loaded
         // #showRecordingTools may not exist, so the change handler has nowhere to attach.
         // It therefore doesn't work until we change pages.
-        $('#accordion').on('change', '#showRecordingTools', function() {
+        $('#toolbox').on('change', '#showRecordingTools', function() {
             var page: HTMLIFrameElement = <HTMLIFrameElement>parent.window.document.getElementById('page');
             if (!page) return; // unit testing?
             if (this.checked) {