+++ src/BloomExe/Edit/LeveledReaderTool.cs
@@ -8,7 +8,7 @@ namespace Bloom.Edit
 {
 	class LeveledReaderTool : ToolboxTool
 	{
-		public const string ToolId = "leveledReader";  // Avoid changing value; see AccordionToo.JsonToolId
+		public const string ToolId = "leveledReader";  // Avoid changing value; see ToolboxTool.JsonToolId
 		public override string JsonToolId { get { return ToolId; } }
 	}
 }